requirements.txt
# pyup: ignore file
# This file is autogenerated. Do not edit it manually.
# Run `make freeze-requirements` to update requirements.txt
# with package version changes made in requirements-app.txt
apig-wsgi==2.15.0
boto==2.49.0
cffi==1.15.1
celery[sqs]==5.2.7
docopt==0.6.2
environs==9.3.2 # pyup: <9.3.3 # marshmallow v3 throws errors
fido2==0.9.1
Flask-Bcrypt==1.0.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
git+https://github.com/mitsuhiko/flask-sqlalchemy.git@500e732dd1b975a56ab06a46bd1a20a21e682262#egg=Flask-SQLAlchemy==2.3.2.dev20190108
Flask==2.2.2
click-datetime==0.2
gevent==22.10.2
gunicorn==20.1.0
iso8601==0.1.14
idna==2.10 # pinned to align with test moto dependency requirements (for <=2.9)
jsonschema==3.2.0
marshmallow-sqlalchemy==0.23.1
marshmallow==2.21.0 
python-magic==0.4.22
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2021.3
PyYAML==5.4.1
SQLAlchemy==1.4.37
sentry-sdk[flask]==1.10.1
cachelib==0.9.0
newrelic==6.10.0.165
notifications-python-client==6.4.1
python-dotenv==0.21.0
pwnedpasswords==2.0.0
tldextract==3.4.0
nanoid==2.0.0
unidecode==1.3.6
more-itertools==8.14.0
# PaaS
awscli-cwlogs==1.4.6
aws-embedded-metrics==1.0.7
# Putting upgrade on hold due to new version introducing breaking changes
Werkzeug==2.2.2
MarkupSafe==2.1.1
# REVIEW: v2 is using sha512 instead of sha1 by default (in v1)
itsdangerous==2.1.2
git+https://github.com/cds-snc/notifier-utils.git@50.1.0#egg=notifications-utils

# rsa==4.9  # awscli 1.22.38 depends on rsa<4.8
typing-extensions==4.4.0
greenlet==2.0.1

## The following requirements were added by pip freeze:
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.2.0
awscli==1.27.10
bcrypt==4.0.1
billiard==3.6.4.0
bleach==3.3.1
blinker==1.5
boto3==1.26.10
botocore==1.29.10
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==39.0.0
docutils==0.16
filelock==3.9.0
flask-redis==0.4.0
frozenlist==1.3.3
future==0.18.2
importlib-metadata==6.0.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
Mako==1.2.4
mistune==0.8.4
multidict==6.0.4
ordered-set==4.1.0
packaging==23.0
phonenumbers==8.13.0
prompt-toolkit==3.0.36
py-w3c==0.3.1
pyasn1==0.4.8
pycparser==2.21
pycurl==7.44.1
PyPDF2==1.28.6
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
redis==4.4.1
requests==2.28.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
smartypants==2.0.1
statsd==3.3.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.8.2
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
